% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateFeatures.R
\name{add_calendar_variables}
\alias{add_calendar_variables}
\title{Add Calendar Features}
\usage{
add_calendar_variables(dt, datetimecol, UKHolidays = NULL)
}
\arguments{
\item{dt}{\code{data.table} to have date/time feature columns added}

\item{datetimecol}{column name of \code{POSIX} time stamps}

\item{UKHolidays}{optional table of holidays from \code{read_holiday_ics}}
}
\value{
Table of UK Holidays
}
\description{
Add Calendar Features
}
\details{
Details go here...
}
\keyword{Date}
\keyword{Features}
