% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRfunctions.R
\name{cov.cor_matrix}
\alias{cov.cor_matrix}
\title{Multivariate Gaussian covariance/correlation matrices}
\usage{
cov.cor_matrix(u_data, kfold = NULL, cov_cor = "covariance", ...)
}
\arguments{
\item{u_data}{A dataframe of uniform distributed variables.}

\item{kfold}{A vector of kfold identifiers.}

\item{cov_cor}{specify either covariance or correlation}
}
\value{
A list of covariance/correlation matrices corresponding to kfold ids
}
\description{
This function produces a list of Multivariate Gaussian spatial/tempral/spatiotemporal covarance/correlation matrices from PIT transformed variables
}
\details{
Details go here...
}
