% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcor_matrix.R
\name{covcor_matrix}
\alias{covcor_matrix}
\title{Multivariate Gaussian covariance/correlation matrices}
\usage{
covcor_matrix(
  u_data,
  kfold = NULL,
  cov_cor = "covariance",
  use = "pairwise.complete.obs",
  boundary_threshold = NA,
  forcePD = F,
  scale = F,
  ...
)
}
\arguments{
\item{u_data}{A dataframe of standard uniform distributed variables.}

\item{kfold}{A vector of kfold identifiers.}

\item{cov_cor}{specify either covariance or correlation.}

\item{boundary_threshold}{Data outside [0,1] are set to \code{boundary_threshold} or 
\code{1-boundary_threshold}. Set \code{boundary_threshold=NA} to
exclude data outside [0,1].}

\item{forcePD}{Use \code{nearPD()} to return a positive definite cov/cor matrix}

\item{scale}{Scale gaussian variable inputs to have unit variance.
for covariance matrix? Defaults to FALSE.}
}
\value{
A list of covariance/correlation matrices corresponding to unique values of \code{kfold}
}
\description{
This function produces a list of Multivariate Gaussian spatial/tempral/spatiotemporal
covarance/correlation matrices from uniform distributed variables.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}; Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
