% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exclude_fun}
\alias{exclude_fun}
\title{set up exclude index for modelling fitting}
\usage{
exclude_fun(data, exclude_train = NULL)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory
variables. May optionally contain a column of exclude flags named by \code{exclude_train}.}

\item{exclude_train}{control for exclusion of rows in data for the model training only, with various options, either:
\itemize{
 \item the column name of the binary/boolean exclude flag supplied in data.
 \item a vector of binary/boolean exclusion flags of length nrow(data)
 \item NULL indicates no exclusion
}}
}
\value{
A vector of binary flags for removing training data points
}
\description{
This generates a vector of length nrow(data) to flag removal from model **training**
}
\details{
....
}
\author{
Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
