% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamboostLSS_2_MultiQR.R
\name{gamboostLSS_2_MultiQR}
\alias{gamboostLSS_2_MultiQR}
\title{Generate predictions from \code{gamboostLSS} object}
\usage{
gamboostLSS_2_MultiQR(
  data,
  models,
  quantiles = seq(0.05, 0.95, by = 0.05),
  params = F
)
}
\arguments{
\item{data}{A \code{data.frame} containing explanatory variables.}

\item{models}{A Para_gamboostLSS object.}

\item{quantiles}{Vector of quantiles to be calculated}

\item{params, }{return distribution parameters instead?}
}
\value{
A \code{MultiQR} object derived from gamlss predictive distributions. Alternatively, a matrix containing the parameters of the predictive gamboostLSS distributions.
}
\description{
Convert gamboostLSS objects to MultiQR or alternatively return
predicted parameters of predictive distribution.

Note that this function may be superseded by an S3 method in future versions of
\code{ProbCast}.

Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
