% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRfunctions.R
\name{gamboostLSS_2_PIT}
\alias{gamboostLSS_2_PIT}
\title{Probability integral transform for Para_gamboostLSS objects}
\usage{
gamboostLSS_2_PIT(models, data, dist_fun, response_name, ...)
}
\arguments{
\item{models}{A Para_gamboostLSS object.}

\item{data}{Input data corresponding to \code{qrdata}.}

\item{dist_fun}{cumulative distribution function corresponging to families specified in gamboostLSS model (see example).}

\item{response_name}{name of response variable in \code{data} object.}
}
\value{
The probability integral transform of \code{data} through the predictive distribution defined by a list of gamboostLSS objects.
}
\description{
This function produces a fan plot of a MultiQR object.
}
\details{
Details go here...
}
