% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_gbm.R
\name{mqr_qreg_gbm}
\alias{mqr_qreg_gbm}
\title{Multiple Quantile Regression Using Gradient Boosted Decision Trees}
\usage{
mqr_qreg_gbm(
  data,
  formula,
  quantiles = c(0.25, 0.5, 0.75),
  cv_folds = NULL,
  perf.plot = F,
  pred_ntree = NULL,
  cores = 1,
  pckgs = NULL,
  sort = T,
  sort_limits = NULL,
  save_models_path = NULL,
  return_op_model = FALSE,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory
variables. May optionally contain a column called "kfold" with
numbered/labeled folds and "Test" for test data.}

\item{quantiles}{The quantiles to fit models for.}

\item{cv_folds}{Control for cross-validation if not supplied in \code{data}.}

\item{perf.plot}{Plot GBM performance?}

\item{pred_ntree}{predict using a user-specified tree.
If unspecified an out-of-the bag estimate will be used unless internal
gbm cross-validation folds are specified in \code{...}.}

\item{cores}{the number of available cores. Defaults to one, i.e. no parallelisation, although in this case the user
must still specify \code{pckgs} if applicable.}

\item{pckgs}{specify additional packages required for
each worker (e.g. c("data.table") if data stored as such).}

\item{sort}{Sort quantiles using \code{SortQuantiles()}?}

\item{sort_limits}{\code{Limits} argument to be passed to \code{SortQuantiles()}. Constrains quantiles to upper and 
lower limits given by \code{list(U=upperlim,L=lowerlim)}.}

\item{save_models_path}{Path to save models. Model details and file extension pasted onto this string. 
Defaults to \code{NULL}, i.e. no model save.}

\item{return_op_model}{return an operational model trained on \strong{all available data} for prediction? see \code{predict.pc_gbm()}.}

\item{...}{Additional arguments passed to \code{gbm()}.}

\item{formala}{A \code{formula} object with the response on the left
of an ~ operator, and the terms, separated by + operators, on the right}
}
\value{
Quantile forecasts in a \code{MultiQR} object. Optionally a list containing a \code{MultiQR}
object and a model for operational prediction
}
\description{
This function fits multiple boosted quantile regression trees 
using \code{gbm} with facilities for cross-validation.
}
\details{
The returned predictive quantiles are those produced out-of-sample for each
cross-validation fold (using models trained on the remaining folds but not "Test" data).
Predictive quantiles corresponding to "Test" data are produced using models trained on all
non-test data. 

When \code{return_op_model = TRUE} the function returns a list containing the \code{MultiQR} object
and an operational model for generating predictions.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}; Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
