% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_qgam.R
\name{MQR_qreg_qgam}
\alias{MQR_qreg_qgam}
\title{Multiple Quantile Regression Using qgam}
\usage{
MQR_qreg_qgam(
  data,
  formula,
  quantiles = c(0.25, 0.5, 0.75),
  CVfolds = NULL,
  ...,
  argGam = list(),
  w = rep(1, nrow(data)),
  parallel = F,
  cores = NULL,
  pckgs = NULL,
  Sort = T,
  SortLimits = NULL,
  save_models_path = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory variables. May optionally contain a collumn called "kfold" with numbered/labeled folds and "Test" for test data.}

\item{formula}{A \code{formula} object with the response on the left of an ~ operator, and the terms, separated by + operators, on the right}

\item{quantiles}{The quantiles to fit models for.}

\item{CVfolds}{Control for cross-validation if not supplied in \code{data}.}

\item{argGam}{List of parameters to be passed to internal call to \code{mgcv::gam}. See \code{?qgam} for details.}

\item{parallel}{\code{boolean} parallelize cross-validation process?}

\item{cores}{if parallel is TRUE then number of available cores}

\item{pckgs}{if parallel is TRUE then  specify packages required for each worker (e.g. c("data.table) if data stored as such)}

\item{Sort}{\code{boolean} Sort quantiles using \code{SortQuantiles()}?}

\item{SortLimits}{\code{Limits} argument to be passed to \code{SortQuantiles()}. Constrains quantiles to upper and lower limits given by \code{list(U=upperlim,L=lowerlim)}.}

\item{save_models_path}{Path to save models. Model details and file extension pasted onto this string.}
}
\value{
Quantile forecasts in a \code{MultiQR} object.
}
\description{
This function fits multiple quantile generalized additive models (QGAMs) with facilities for cross-validation.
}
\details{
Actual fitting happens in \code{qgam::qgam}.
}
\keyword{Quantile}
\keyword{Regression}
