% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Para_gamboostLSS.R
\name{Para_gamboostLSS}
\alias{Para_gamboostLSS}
\title{Fit boosted \code{gamlss}-type semi-parametric models (depreciated)}
\usage{
Para_gamboostLSS(
  data,
  formula,
  families = gamboostLSS::GaussianLSS(),
  parallel = F,
  cores = NULL,
  pckgs = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory variables.
May optionally contain a column called "kfold" with numbered/labeled
folds and "Test" for test data.}

\item{formula}{A formula or list of formulas for differences
between formulas for location, scale, shape etc..(see \code{?gamboostLSS})}

\item{families}{A gamboosLSS family object, which is used to define the
distribution and the link functions of the various parameters.}

\item{parallel}{\code{boolean} parallelize cross-validation process?
Parallelisation is over cross-validation folds.}

\item{pckgs}{if parallel is TRUE then  specify packages required for
each worker (e.g. c("data.table) if data stored as such)}

\item{...}{Additional arguments passed to \code{gamboostLSS()}.}
}
\value{
A list of \code{gamboostlss} objects. Each list element
corresponds to a cross-validation fold and contains a \code{gamlss} model
trained on all other folds.
}
\description{
This function is now depreciated and may be removed in future versions of this package.
Use \code{ppd_gamboostlss()} instead.
}
\author{
Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
