% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIT.R
\name{PIT.gamboostLSS}
\alias{PIT.gamboostLSS}
\title{Probability integral transform for Para_gamboostLSS objects}
\usage{
\method{PIT}{gamboostLSS}(models, data, dist_fun, response_name, ...)
}
\arguments{
\item{models}{A Para_gamboostLSS object.}

\item{data}{Input data corresponding to \code{qrdata}.}

\item{dist_fun}{cumulative distribution function corresponging to families specified in gamboostLSS model (see example).}

\item{response_name}{name of response variable in \code{data} object.}
}
\value{
The probability integral transform of \code{data} through the predictive distribution defined by a list of gamboostLSS objects.
}
\description{
This function calculates the probability integral transformation
for \code{gamboostLSS} models.
}
\author{
Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
