% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRfunctions.R
\name{PIT.PPD}
\alias{PIT.PPD}
\title{Probability integral transform: S3 Method for for PPD}
\usage{
\method{PIT}{PPD}(ppd, data, inverse = FALSE, inv_probs, ...)
}
\arguments{
\item{ppd}{A \code{PPD} object.}

\item{data}{Input data corresponding to \code{qrdata}.}

\item{...}{Additional arguments passed to \code{condCDF}.}
}
\value{
The probability integral transform of \code{data} through the predictive distribution defined by \code{ppd}, a list of gamlss objects.
}
\description{
This function produces a fan plot of a MultiQR object.
}
\details{
Details go here...
}
