% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPD_2_MultiQR.R
\name{PPD_2_MultiQR}
\alias{PPD_2_MultiQR}
\title{Convert PPD to MultiQR, or alternatively return predicted parameters of predictive distribution.}
\usage{
PPD_2_MultiQR(data, models, quantiles = seq(0.05, 0.95, by = 0.05), params = F)
}
\arguments{
\item{data}{A \code{data.frame} containing explanatory variables.}

\item{models}{An \code{PPD} object.}

\item{quantiles}{Vector of quantiles to be calculated}

\item{params}{return distribution parameter predictions?}
}
\value{
A \code{MultiQR} object derived from gamlss predictive distributions. Alternatively, a matrix condaining the parameters of the predictive  gamlss distributions.
}
\description{
Convert PPD to MultiQR, or alternatively return predicted parameters of predictive distribution.
}
\details{
Details go here...
}
