% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Para_gamboostLSS.R
\name{ppd_gamboostlss}
\alias{ppd_gamboostlss}
\title{Fit boosted \code{gamlss}-type semi-parametric models}
\usage{
ppd_gamboostlss(
  data,
  formula,
  cv_folds = NULL,
  family = gamboostLSS::GaussianLSS(),
  cores = 1,
  pckgs = NULL,
  save_models_path = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory variables.
May optionally contain a column called "kfold" with numbered/labeled
folds and "Test" for test data.}

\item{formula}{A formula or named list of formulas 
for the location, scale, shape etc..(see \code{?gamboostLSS})}

\item{cv_folds}{Control for cross-validation if not supplied in \code{data}.}

\item{family}{A gamboosLSS family object, which is used to define the
distribution and the link functions of the various parameters.}

\item{cores}{the number of available cores. Defaults to one, i.e. no parallelisation, although in this case the user
must still specify \code{pckgs} if applicable.}

\item{pckgs}{specify additional packages required for
each worker (e.g. c("data.table") if data stored as such).}

\item{save_models_path}{Path to save models. Model details and file extension pasted onto this string.}

\item{...}{additional arguments passed to \code{gamboostLSS()}, 
e.g. use control = mboost::boost_control() to specify boosting steps, shrinkage etc.}
}
\value{
A list of \code{gamboostlss} objects. Each list element
corresponds to a cross-validation fold and contains a \code{gamlss} model
trained on all other folds.
}
\description{
This function is a wrapper for the function \code{gamboostLSS}, which
fits semi-parametric regression models for predictive distributions with up to
four parameters (location, scale, shape1, shape2) via gradient boosting.
}
\author{
Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
