% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_lightgbm.R
\name{predict.qreg_lightgbm}
\alias{predict.qreg_lightgbm}
\title{Predict method for Multiple Quantile Regression from lightgbm Gradient Boosted Decision Trees.}
\usage{
\method{predict}{qreg_lightgbm}(
  object,
  newdata = NULL,
  quantiles = NULL,
  model_name = NULL,
  sort = T,
  sort_limits = NULL,
  ...
)
}
\arguments{
\item{object}{fitted model of class \code{qreg_lightgbm} obtained from the function \code{qreg_lightgbm()}.}

\item{newdata}{data.frame of observations for which to make predictions}

\item{quantiles}{the quantiles to predict. Default is all the quantiles present in \code{object}}

\item{model_name}{the name of the model to be used for prediction.
Unless specified, \code{object$default_model} is used.}

\item{sort}{sort quantiles using \code{SortQuantiles()}?}

\item{sort_limits}{\code{Limits} argument to be passed to \code{SortQuantiles()}. Constrains quantiles to upper and
lower limits given by \code{list(U=upperlim,L=lowerlim)}.}

\item{...}{additional arguments; not currently used.}
}
\value{
Quantile forecasts in a \code{MultiQR} object.
}
\description{
This function returns multiple quantile predictions as an object of class \code{MultiQR}
based on a ProbCast gbm fitted model. S3 Method for for \code{qreg_lightgbm} objects
}
\details{
this function returns predictive quantiles for each row in \code{newdata},
the result is returned as a \code{MultiQR} object
}
\author{
Gordon McFadzean, \email{gordon.mcfadzean@tneigroup.com}
}
\keyword{Quantile}
\keyword{Regression}
