% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_gam.R
\name{print.qreg_gam}
\alias{print.qreg_gam}
\title{Print \code{qreg_gbm} object}
\usage{
\method{print}{qreg_gam}(x, ...)
}
\arguments{
\item{x}{object of class \code{qreg_gam} obtained from the function \code{qreg_gbm()}.}
}
\description{
Print \code{qreg_gbm} object
}
\details{
this function prints details of the call to \code{qreg_gam()},
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
