% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_gbm.R
\name{print.qreg_gbm}
\alias{print.qreg_gbm}
\title{Print \code{qreg_gbm} object}
\usage{
\method{print}{qreg_gbm}(x, ...)
}
\arguments{
\item{x}{object of class \code{qreg_gbm} obtained from the function \code{qreg_gbm()}.}

\item{...}{additional arguments; not currently used.}
}
\description{
Print \code{qreg_gbm} object
}
\details{
this function prints details of the call to \code{qreg_gbm()},
}
\author{
Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
