% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_lightgbm.R
\name{print.qreg_lightgbm}
\alias{print.qreg_lightgbm}
\title{Print \code{qreg_lightgbm} object}
\usage{
\method{print}{qreg_lightgbm}(x, ...)
}
\arguments{
\item{x}{object of class \code{qreg_lightgbm} obtained from the function \code{qreg_lightgbm()}.}

\item{...}{additional arguments; not currently used.}
}
\description{
Print \code{qreg_lightgbm} object
}
\details{
this function prints details of the call to \code{qreg_lightgbm()},
}
\author{
Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
