% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_mboost.R
\name{print.qreg_mboost}
\alias{print.qreg_mboost}
\title{Print \code{qreg_mboost} object}
\usage{
\method{print}{qreg_mboost}(x, ...)
}
\arguments{
\item{x}{object of class \code{qreg_mboost} obtained from the function \code{qreg_mboost()}.}

\item{...}{additional arguments; not currently used.}
}
\description{
Print \code{qreg_mboost} object
}
\details{
this function prints details of the call to \code{qreg_mboost()},
}
\author{
Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
