% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_gam.R
\name{qreg_gam.add_quantiles}
\alias{qreg_gam.add_quantiles}
\title{Add new qunatile regression models and predictions to \code{qreg_gam} model.}
\usage{
qreg_gam.add_quantiles(object, data, quantiles)
}
\arguments{
\item{object}{An \code{qreg_gam} object.}

\item{data}{The data used to fit \code{object}}

\item{quantiles}{The new quantiles to be added to \code{object}. Quantiles already
in \code{object} will not be re-fit and a warning thrown.}
}
\value{
An updated \code{qreg_gam}.
}
\description{
This function adds new conditional quantile models and associated predictions
to an exisiting \code{qreg_gam} object. This is a much faster approach to add
new quantiles than re-estimating the main GAM using \code{qreg_gam()}. This
function is also called within \code{qreg_gam()}.
}
\details{
Fits new qunatiles to \code{qreg_gam} object.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
