% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_gam.R
\name{qreg_gam}
\alias{qreg_gam}
\title{Multiple Quantile Regression Using Generalised Additive Models and Linear Quantile Regression}
\usage{
qreg_gam(
  data,
  formula,
  formula_qr = NULL,
  model_res2 = F,
  formula_res2 = NULL,
  quantiles = c(0.25, 0.5, 0.75),
  cv_folds = NULL,
  use_bam = T,
  exclude_train = NULL,
  w = NULL,
  sort = T,
  sort_limits = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory
variables. May optionally contain a collumn called "kfold" with
numbered/labeled folds and "Test" for test data. If \code{data} contains a column called
"gam_pred" then gam modelling will be skipped and this will be used in quantile regression.}

\item{formula_qr}{Formula for linear quantile regression model for GAM residuals. Term \code{gam_pred}
is the prediction from the above GAM may be included in this formula. If null, the "terms" of the GAM model
are used as features in linear quantile regression.}

\item{model_res2}{If \code{TRUE} also model squared residuals of GAM using a GAM. Defaults to \code{FALSE}.}

\item{formula_res2}{Formula for GAM to predict squared residuals.}

\item{quantiles}{The quantiles to fit models for.}

\item{cv_folds}{Control for cross-validation with various options, either:
\itemize{
 \item the column name of the fold index supplied in data. Observations and inputs 
  in the index labeled "Test" will serve as test data and held out in model training.
 \item an integer giving the number of cross validation folds to generate. Folds are constructed as block chunks. 
 Default behaviour is 5 folds.
 \item NULL indicates that no cross validation should be performed and the returned model is trained on all \code{data}.
}}

\item{use_bam}{If \code{TRUE} (default) then GAM is fit using (\code{bam()}) in stead
of \code{gam()}. \code{bam} is better suited to large datasets but not all
\code{gam} model options are available with \code{bam}. Alternative smooths, such as
cubic regression splines aid faster estimation than the default smooth \code{bs="tp"}.
See \code{bam()} documentation for further details.}

\item{exclude_train}{A column name in \code{data} indicating if a row should be excluded from model
training, i.e. if it contains bad data (will be coerced to \code{logical}). Alterntively,
an \code{integer} or \cpde{logical} vector with length equal to the number of rows in \code{data} indicating
the same. Rows labeled \code{TRUE} are excluded from model training.}

\item{w}{Weights on the contribution of the data to log likelihood in \code{gam/bam} models ONLY (and not quantile loss for quantile regressions). See \code{?gam, ?bam} for more details.}

\item{sort}{\code{boolean} Sort quantiles using \code{SortQuantiles()}?}

\item{sort_limits}{\code{Limits} argument to be passed to \code{SortQuantiles()}. Constrains quantiles to upper and lower limits given by \code{list(U=upperlim,L=lowerlim)}.}

\item{...}{Additional agruments passter to \code{gam()} (or \code{bam()}).}

\item{formala}{A \code{formula} object with the response on the left
of an ~ operator, and the terms, separated by + operators, on the right passed to \code{gam()}
or \code{bam()} from \code{mgcv}.}
}
\value{
Returns a \code{list} containing predictive quantiles (in a \code{MultiQR} object), GAM models, and deterministic predictions
from GAMs.
}
\description{
This function fits multiple conditional linear quantile regression models to the residuals of
a generalised additive model using \code{mgcv} and with facilities for cross-validation.
Quantile regression may be performed using user-specified formula or the design matrix of
the fitted GAM.
}
\details{
The returned predictive quantiles and GAM predictions are those produced out-of-sample for each
cross-validation fold (using models trained on the remaining folds but not "Test" data).
Predictive quantiles corresponding to "Test" data are produced using models trained on all
non-test data.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
