% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_gam.R
\name{qreg_gam.update}
\alias{qreg_gam.update}
\title{Predict from model based on Generalised Additive Model and Linear Quantile Regression}
\usage{
qreg_gam.update(object, newdata, model_name = NULL)
}
\arguments{
\item{object}{An \code{qreg_gam} object containing the model to update.}

\item{newdata}{A data frame or data table containing the values of the model
covariates and target variable required to update the model.}

\item{model_name}{The name of the model in \code{object} to be updated.}
}
\value{
An updated \code{qreg_gam} model.
}
\description{
This function predicts from multiple conditional linear quantile regression models of the residuals of
a generalised additive model fit using \code{qreg_gam}.
}
\details{
Update the \code{bam} component of an \code{qreg_gam} model.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
