% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_mboost.R
\name{qreg_mboost}
\alias{qreg_mboost}
\title{Multiple Quantile Regression using \code{mboost}}
\usage{
qreg_mboost(
  data,
  formula,
  quantiles = c(0.25, 0.5, 0.75),
  cv_folds = NULL,
  w = rep(1, nrow(data)),
  cores = 1,
  pckgs = NULL,
  sort = T,
  sort_limits = NULL,
  save_models_path = NULL,
  only_mqr = FALSE,
  exclude_train = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory variables.
May optionally contain a column called "kfold" with numbered/labeled folds
and "Test" for test data.}

\item{quantiles}{A vector with length>=2 containing the quantiles to fit models for.}

\item{cv_folds}{Control for cross-validation if not supplied in \code{data}.}

\item{w}{an optional numeric vector of weights to be used in the fitting process.}

\item{cores}{the number of available cores. Defaults to one, i.e. no parallelisation, although in 
this case the user must still specify \code{pckgs} if applicable.}

\item{pckgs}{specify additional packages required for
each worker (e.g. c("data.table") if data stored as such).}

\item{sort}{Sort quantiles using \code{SortQuantiles()}?}

\item{sort_limits}{\code{Limits} argument to be passed to \code{SortQuantiles()}. Constrains quantiles 
to upper and lower limits given by \code{list(U=upperlim,L=lowerlim)}.}

\item{save_models_path}{Path to save models. Model details and file extension pasted onto this string.
Defaults to \code{NULL}, i.e. no model save.}

\item{...}{extra hyper-parameters to be passed to \code{mboost()}.
e.g. use \code{control = mboost::boost_control()} to specify boosting steps and shrinkage.}

\item{formaula}{A \code{formula} object with the response on the left
of an ~ operator, and the terms, separated by + operators, on the right.}
}
\value{
Quantile forecasts in a \code{MultiQR} object.
}
\description{
This function fits multiple quantile regression models using \code{mboost}, with
facilities for cross-validation. \code{mboost} accommodates both generalised
additive models, decision trees and other learners. See \code{?mboost} for more
details.
}
\details{
The returned predictive quantiles are those produced out-of-sample for each
cross-validation fold (using models trained on the remaining folds but not "Test" data).
Predictive quantiles corresponding to "Test" data are produced using models trained on all
non-test data.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}; Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
