% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateFeatures.R
\name{read_holiday_ics}
\alias{read_holiday_ics}
\title{Read UK Holidays}
\usage{
read_holiday_ics(folder = getwd())
}
\arguments{
\item{folder}{Folder location of .ics file(s) of UK holidays from https://www.gov.uk/bank-holidays}
}
\value{
Table of UK Holidays
}
\description{
Read UK Holidays
}
\details{
Details go here...
}
\keyword{Date}
\keyword{Features}
