% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Reliability Diagram for MultiQR}
\usage{
reliability(
  qrdata,
  realisations,
  kfolds = NULL,
  subsets = NULL,
  breaks = 4,
  bootstrap = NULL,
  plot.it = T,
  ...
)
}
\arguments{
\item{qrdata}{\code{MultiQR} object.}

\item{realisations}{Vector of realisations corresponding to
rows of \code{qrdata}. Missing data as \code{NA}s accepted.}

\item{kfolds}{Optional vector of cross-validation fold labels corresponding
to rows of \code{qrdata}. Cannot be used with \code{subsets}.}

\item{subsets}{Optional vector of covariates to bin data by.
Breaks between bins are the empirical quantiles of
\code{subsets} by default or all unique factors or charater strings. Custom
breaks may be specifed, see \code{breaks}.
Cannot be used with \code{kfolds}.}

\item{breaks}{Either the number of quantiles to use to bin \code{subsets} by (resulting
in \code{breaks+1} bins, defaults to \code{breaks=4}), or, if \code{length(breaks) > 1}, a vector of spcific break
points. Only used if \code{subsets} provided.
points. \code{subsets} must be provided.}

\item{bootstrap}{Number of boostrap samples used to generate 95\% confidence intervals.}

\item{plot.it}{\code{boolean}. Make a plot?}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
Reliability data and, if \code{plot.it=TRUE}, a reliability diagram.
}
\description{
Calculated empirical exceedence of each quantile and 
plots a reliabiltiy diagram for a \code{MultiQR} object.

Optionally, results may be split by cross-validation fold or covariate and/or
confidence intervals may be estimated.
}
\details{
Missing values in \code{realisations} are handled by \code{na.rm=T} when
calculating average exceedence of a given quantile.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
