% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleMQR.R
\name{rescaleMQR}
\alias{rescaleMQR}
\title{Re-scale MultiQR object}
\usage{
rescaleMQR(mqr, mult = 1, add = 0, Order = "ma")
}
\arguments{
\item{mqr}{A \code{MultiQR} object to be re-scaled}

\item{mult}{Multiplicative adjustment}

\item{add}{Additive adjustment}

\item{order}{Order in which to perform adjustment}
}
\value{
A \code{MultiQR} object
}
\description{
This function re-scales MultiQR object.
}
\details{
Details go here...
}
\keyword{Quantile}
\keyword{Regression,}
\keyword{plot}
