% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_lightgbm.R
\name{retrain_all}
\alias{retrain_all}
\title{Model retraining: S3 Generic Method}
\usage{
retrain_all(object, ...)
}
\arguments{
\item{object}{A fitted model object. Currently supported: \code{qreg_lightgbm}.}

\item{...}{Additional arguments.}
}
\value{
A fitted model, periodically retrained over the Test set.
}
\description{
Model retraining: S3 Generic Method
}
\details{
This is an S3 method, see specific methods \code{\link{retrain_all.qreg_lightgbm}}
for details on functionality.
}
\author{
Rosemary Tawn, \email{rosemary.tawn@tneigroup.com}
}
