% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpness.R
\name{sharpness}
\alias{sharpness}
\title{Sharpness/Interval Width for \code{MultiQR} Objects}
\usage{
sharpness(qrdata, kfolds = NULL, bootstrap = NULL, ...)
}
\arguments{
\item{qrdata}{\code{MultiQR} object.}

\item{kfolds}{Optional vector of fold/test labels corresponding to rows of \code{qrdata}.
Cannot be used with \code{subsets}.}

\item{bootstrap}{Number of boostrap samples used to generate 95\% confidence intervals.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
Quantile Score data and, if \code{plot.it=T}, a plot.
}
\description{
This function calculates the interval width for symetric quantiles in a \code{MultiQR}
object. Optionally, results are produced by cross-validation fold or covariate,
95\% confidence intervals are estimated via bootstrap, and results are plotted.
}
\details{
Missing values in \code{realisations} are handled by \code{na.rm=T} when
calculating average exceedence of a given quantile.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
