% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRfunctions.R
\name{SortQuantiles}
\alias{SortQuantiles}
\title{Sort multiple quantiles stored in a data.frame/MultiQR object}
\usage{
SortQuantiles(data, Limits = NULL)
}
\arguments{
\item{data}{A \code{MultiQR, data.frame} containing quantile forecasts}

\item{Limits}{List Upper and lower limits to apply to quantiles. E.g. list(U=0.999,L=0.001)}
}
\value{
A \code{MultiQR, data.frame} object with ordered quanitles.
}
\description{
This function sorts quantiles so that q1<q2<q3<... is satisfied.
}
\details{
Details go here...
}
\keyword{Quantile}
\keyword{Regression}
