% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_gam.R
\name{summary.qreg_gam}
\alias{summary.qreg_gam}
\title{Summary \code{qreg_gbm} object}
\usage{
\method{summary}{qreg_gbm}(x, ...)
}
\arguments{
\item{x}{object of class \code{qreg_gam} obtained from the function \code{qreg_gbm()}.}
}
\description{
Summary \code{qreg_gbm} object
}
\details{
this function gives a more detailed summary of, \code{x}, a \code{qreg_gam} object,
than \code{print()}
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
