% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_lightgbm.R
\name{summary.qreg_lightgbm}
\alias{summary.qreg_lightgbm}
\title{Summary \code{qreg_lightgbm} object}
\usage{
\method{summary}{qreg_lightgbm}(x, ...)
}
\arguments{
\item{x}{object of class \code{qreg_lightgbm} obtained from the function \code{qreg_lightgbm()}.}

\item{...}{additional arguments; not currently used.}
}
\description{
Summary \code{qreg_lightgbm} object
}
\details{
this function gives a more detailed description details of the \code{x} object,
than \code{print()}
}
\author{
Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
