% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_mboost.R
\name{summary.qreg_mboost}
\alias{summary.qreg_mboost}
\title{Summary \code{qreg_mboost} object}
\usage{
\method{summary}{qreg_mboost}(x, ...)
}
\arguments{
\item{x}{object of class \code{qreg_mboost} obtained from the function \code{qreg_mboost()}.}

\item{...}{additional arguments; not currently used.}
}
\description{
Summary \code{qreg_mboost} object
}
\details{
this function gives a more detailed description details of the \code{x} object,
than \code{print()}
}
\author{
Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
