% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tails_ev.R
\name{tails_ev}
\alias{tails_ev}
\title{Extreme value distirbutions for distribution tails}
\usage{
tails_ev(
  data,
  mqr_data,
  tail_starts = range(as.numeric(gsub("q", "", names(mqr_data))), na.rm = T),
  formula,
  formula_r = formula,
  CVfolds = NULL,
  BadData_col = NULL,
  evgam_family = "gpd",
  print_summary = F,
  return_models = F
)
}
\arguments{
\item{data}{The modelling table as a \code{data.frame} or \code{data.table}}

\item{mqr_data}{A \code{MultiQR} object corresponding to \code{data}}

\item{tail_starts}{The upper and lower probability levels beyond which the
tail distribution is used. E.g. \code{c(5,95)} to use tail distribution for probabilities
below quantile 0.05 and above 0.95. Defaults to min and max from \code{mqr_data}}

\item{formula_r}{as \code{formula} but for the upper tail only.}

\item{CVfolds}{Control for cross-validation if not supplied in \code{data}.}

\item{BadData_col}{Name of a boolean column in \code{data} indicating bad data
which should be excluded from parameter estimation.}

\item{evgam_family}{specifies distribution, see \code{?evgam}.}

\item{print_summary}{If \code{TRUE}, summary is printed for each evgam fit.}

\item{return_models}{If \code{TRUE}, the function returns a list including both the data and the fitted data models.}

\item{formala}{A \code{formula} object with the response on the left
of an ~ operator, and the terms, separated by + operators, on the right. This formula is
used for both upper and lower tails unless \code{formula_r} is specified in which case
\code{formula} is used for the lower tail only.}
}
\value{
If `return_models` is \code{TRUE}, then returns a list containing a list of tail models, and \code{data} with
additional columns containing the predicted parameters of the specified tail distribution. Otherwise, just returns \code{data}.
}
\description{
This function fits semi-parametric models for extreme value distributions
using \code{evgam} with facilities for cross-validation.
}
\details{
The returned predictive quantiles are those produced out-of-sample for each
cross-validation fold (using models trained on the remaining folds but not "Test" data).
Predictive quantiles corresponding to "Test" data are produced using models trained on all
non-test data.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
\keyword{Distribution;}
\keyword{Extreme}
\keyword{Tails}
\keyword{Value}
