% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_lightgbm.R
\name{update.qreg_lightgbm}
\alias{update.qreg_lightgbm}
\title{Model update (retraining) method for Multiple Quantile Regression from lightgbm Gradient Boosted Decision Trees.}
\usage{
update.qreg_lightgbm(
  object,
  newdata,
  model_name = NULL,
  pckgs = NULL,
  cores = 1
)
}
\arguments{
\item{object}{fitted model of class \code{qreg_lightgbm} obtained from the function \code{qreg_lightgbm()}.}

\item{newdata}{data.frame of new observations for which to update the model fit}

\item{model_name}{the name of the model to be used for prediction.
Unless specified, \code{object$default_model} is used.}

\item{pckgs}{specify additional packages required for
each worker (e.g. c("data.table") if data stored as such).}

\item{cores}{the number of available cores. Defaults to one, i.e. no parallelisation, although in this case the user
must still specify \code{pckgs} if applicable.}

\item{...}{additional arguments; not currently used.}
}
\value{
a named list containing fitted models as a list of \code{qreg_lightgbm} objects.
}
\description{
This function returns updated fitted \code{qreg_lightgbm} models, which are the 
supplied previously fitted models updated with new data. S3 Method for \code{qreg_lightgbm} objects
}
\details{
this function updates the fitted lightgbm model with the supplied new data. 
NOTE predictions saved within the lightgbm object are not updated within this function.
}
\author{
Gordon McFadzean, \email{gordon.mcfadzean@tneigroup.com}
}
\keyword{Quantile}
\keyword{Regression}
