% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contCDF.R
\name{contCDF}
\alias{contCDF}
\title{Continuous CFDs}
\usage{
contCDF(
  quantiles,
  kfold = NULL,
  inverse = F,
  method = list(name = "spline", splinemethod = "monoH.FC"),
  tails = list(method = "interpolate", L = 0, U = 1),
  ...
)
}
\arguments{
\item{quantiles}{A single-row \code{MultiQR} object.}

\item{method}{Method of interpolation. If \code{method="linear"} linear interpolation is used between quantiles. For spline interpolation, \code{method=list(name=spline,splinemethod)}, where spline method is passed to \code{splinefun}.}

\item{tails}{Method for tails, this must be defined as a list with other control parameters. Default is linear interpolation of tails when \code{method="interpolate"} with boundaries at \code{L=0} and \code{U=1}. The number of interpolation points for the tail can be defined via \code{ntailpoints}. Alternative methods include:
\itemize{
  \item Exponential tails can be specified through \code{method="exponential"}, the user can either supply user defined thickness parameters for the tail via \code{thicknessPL} and \code{thicknessPR}, otherwise a symetrical tail thickness can be defined data-driven by specifying: number of bins \code{nbins}, a MQR object \code{preds}, and the target variable, \code{targetvar}.  
  \item Dynamic exponential tails can be specified through \code{method="dyn_exponential"}, where the tail depends on the values for the upper and lower quantile of \code{qrdata}, these are only valid for an input variable scale of \code{[0,1]}.
  \item The Generalised Pareto Distribution may be chosen with \code{method="gpd"}, and additional arguments \code{scale_r,shape_r,scale_l,shape_l,tail_qs} specifying the right and left scale and shapre parameters and the incraments beyond the final quantiles to evaluate the gpd at, e.g. \code{tail_qs=seq(0.1,1,by=0.1)} to exctend into the tails by +/- 1 in steps of 0.1.
}}

\item{...}{extra arguments to \code{approxfun} or \code{splinefun}. For example to deal with ties if present.}

\item{kfolds}{Fold/test label corresponding to \code{quantiles}.}
}
\value{
A cumulative density function
}
\description{
This function generats a function that represents a smooth CDF for each row of a MultiQR object.
}
\details{
Details go here...
}
