% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRfunctions.R
\name{contCDF}
\alias{contCDF}
\title{Continuous CFDs}
\usage{
contCDF(quantiles, kfold = NULL, inverse = F, method = list(name =
  "spline", splinemethod = "monoH.FC"), tails = list(method =
  "interpolate", L = 0, U = 1))
}
\arguments{
\item{quantiles}{A single-row \code{MultiQR} object.}

\item{method}{Method of interpolation. If \code{method="linear"} linear interpolation is used between quantiles. For spline interpolation, \code{method=list(name=spline,splinemethod)}, where spline method is passed to \code{splinefun}.}

\item{tails}{Method for tails...}

\item{kfolds}{Fold/test label corresponding to \code{quantiles}.}
}
\value{
A cumulative densift function
}
\description{
This function generats a function that represents a smooth CDF for each row of a MultiQR object.
}
\details{
Details go here...
}
