% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcor_matrix.R
\name{covcor_matrix}
\alias{covcor_matrix}
\title{Multivariate Gaussian covariance/correlation matrices}
\usage{
covcor_matrix(u_data, kfold = NULL, cov_cor = "covariance",
  use = "pairwise.complete.obs", boundary_threshold = NA,
  forcePD = F,scale = F, ...)
}
\arguments{
\item{u_data}{A dataframe of uniform distributed variables.}

\item{kfold}{A vector of kfold identifiers.}

\item{cov_cor}{specify either covariance or correlation}

\item{forcePD}{use \code{nearPD} to return a positive definite cov/cor matrix}
\item{scale}{scale gaussian variable inputs to have unit variance for covariance matrix? Defaults to FALSE.}
\item{boundary_threshold}{handling of boundary values. Set to NA to exclude, or small value to impose threshold.}

\item{forcePD}{use \code{nearPD} to return a positive definite cov/cor matrix}
}
\value{
A list of covariance/correlation matrices corresponding to kfold ids
}
\description{
This function produces a list of Multivariate Gaussian spatial/tempral/spatiotemporal covarance/correlation matrices from PIT transformed variables
}
\details{
Details go here...
}
