% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRfunctions.R
\name{gamboostLSS_2_MultiQR}
\alias{gamboostLSS_2_MultiQR}
\title{Convert gamboostLSS objects to MultiQR, or alternatively return predicted parameters of predictive distribution.}
\usage{
gamboostLSS_2_MultiQR(data, models, quantiles = seq(0.05, 0.95, by =
  0.05), params = F)
}
\arguments{
\item{data}{A \code{data.frame} containing explanatory variables.}

\item{models}{A Para_gamboostLSS object.}

\item{quantiles}{Vector of quantiles to be calculated}
}
\value{
A \code{MultiQR} object derived from gamlss predictive distributions. Alternatively, a matrix condaining the parameters of the predictive gamboostLSS distributions.
}
\description{
Convert gamboostLSS objects to MultiQR, or alternatively return predicted parameters of predictive distribution.
}
\details{
Details go here...
}
