% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRfunctions.R
\name{MQR_gbm}
\alias{MQR_gbm}
\title{Multiple Quantile Regression Using GBM}
\usage{
MQR_gbm(data, formula, quantiles = c(0.25, 0.5, 0.75), CVfolds = NULL,
  gbm_params = list(interaction.depth = 4, n.trees = 500, shrinkage =
  0.25, cv.folds = 1, n.minobsinnode = 15, bag.fraction = 0.5, keep.data =
  F), perf.plot = F, parallel = F, pred_ntree = NULL, cores = NULL,
  pckgs = NULL, Sort = T, SortLimits = NULL)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory variables. May optionally contain a collumn called "kfold" with numbered/labeled folds and "Test" for test data.}

\item{quantiles}{The quantiles to fit models for.}

\item{CVfolds}{Control for cross-validation if not supplied in \code{data}.}

\item{gbm_params}{List of parameters to be passed to \code{fit.gbm()}.}

\item{perf.plot}{Plot GBM performance?}

\item{parallel}{\code{boolean} parallelize cross-validation process?}

\item{pred_ntree}{predict using a user-specified tree, if unspecified an out-of-the bag estimate will be used}

\item{cores}{if parallel is TRUE then number of available cores}

\item{pckgs}{if parallel is TRUE then  specify packages required for each worker (e.g. c("data.table) if data stored as such)}

\item{Sort}{\code{boolean} Sort quantiles using \code{SortQuantiles()}?}

\item{SortLimits}{\code{Limits} argument to be passed to \code{SortQuantiles()}. Constrains quantiles to upper and lower limits given by \code{list(U=upperlim,L=lowerlim)}.}

\item{formaul}{A \code{formula} object with the response on the left of an ~ operator, and the terms, separated by + operators, on the right}
}
\value{
Quantile forecasts in a \code{MultiQR} object.
}
\description{
This function fits multiple quantile regreesion GBMs with facilities for cross-validation.
}
\details{
Details go here...
}
\keyword{Quantile}
\keyword{Regression}
