% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_mboost.R
\name{MQR_qreg_mboost}
\alias{MQR_qreg_mboost}
\title{Multiple Quantile Regression Using mboost}
\usage{
MQR_qreg_mboost(data, formula, quantiles = c(0.25, 0.5, 0.75),
  CVfolds = NULL, ..., bc_mstop = 100, bc_nu = 0.1, w = rep(1,
  nrow(data)), parallel = F, cores = NULL, pckgs = NULL, Sort = T,
  SortLimits = NULL)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory variables. May optionally contain a collumn called "kfold" with numbered/labeled folds and "Test" for test data.}

\item{quantiles}{The quantiles to fit models for.}

\item{CVfolds}{Control for cross-validation if not supplied in \code{data}.}

\item{parallel}{\code{boolean} parallelize cross-validation process?}

\item{cores}{if parallel is TRUE then number of available cores}

\item{pckgs}{if parallel is TRUE then  specify packages required for each worker (e.g. c("data.table) if data stored as such)}

\item{Sort}{\code{boolean} Sort quantiles using \code{SortQuantiles()}?}

\item{SortLimits}{\code{Limits} argument to be passed to \code{SortQuantiles()}. Constrains quantiles to upper and lower limits given by \code{list(U=upperlim,L=lowerlim)}.}

\item{formaul}{A \code{formula} object with the response on the left of an ~ operator, and the terms, separated by + operators, on the right}

\item{model_params}{List of parameters to be passed to \code{fit.gbm()}.}
}
\value{
Quantile forecasts in a \code{MultiQR} object.
}
\description{
This function fits multiple quantile regreesion GBMs with facilities for cross-validation.
}
\details{
Details go here...
}
\keyword{Quantile}
\keyword{Regression}
