% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Para_gamboostLSS.R
\name{Para_gamboostLSS}
\alias{Para_gamboostLSS}
\title{Fit a gamboostlss paramertirc forecast model}
\usage{
Para_gamboostLSS(
  data,
  formula,
  families = GaussianLSS(),
  parallel = F,
  cores = NULL,
  pckgs = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory variables. May optionally contain a collumn called "kfold" with numbered/labeled folds and "Test" for test data.}

\item{formula}{A formula or list of formulas for differences between formulas for location, scale, shape etc..(see \code{gamboostLSS})}

\item{families}{A gamboosLSS family object, which is used to define the distribution and the link functions of the various parameters.}

\item{parallel}{\code{boolean} parallelize cross-validation process?}

\item{pckgs}{if parallel is TRUE then  specify packages required for each worker (e.g. c("data.table) if data stored as such)}

\item{...}{Additonal arguments passed to \code{gamboostLSS()}.}
}
\value{
A list of gamboostlss objects corresponsing to kfolds
}
\description{
Fit a gamboostlss paramertirc forecast model
}
