% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Para_gamlss.R
\name{Para_gamlss}
\alias{Para_gamlss}
\title{Fit a gamlss paramertirc forecast model}
\usage{
Para_gamlss(data, formula, sigma.formula = ~1, nu.formula = ~1,
  tau.formula = ~1, family = NO(), parallel = F, cores = NULL,
  pckgs = NULL, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory variables. May optionally contain a collumn called "kfold" with numbered/labeled folds and "Test" for test data.}

\item{formula}{A formula object with the response on the left of an ~ operator, and the terms, separated by + operators, on the right.}

\item{sigma.formula}{A formula object for fitting a model to the sigma parameter, as in the formula above.}

\item{nu.formula}{A formula object for fitting a model to the nu parameter, as in the formula above.}

\item{tau.formula}{A formula object for fitting a model to the tau parameter, as in the formula above.}

\item{family}{A gamlss.family object, which is used to define the distribution and the link functions of the various parameters.}

\item{parallel}{\code{boolean} parallelize model fitting process?}

\item{cores}{if parallel is TRUE then number of available cores}

\item{pckgs}{if parallel is TRUE then  specify packages required for each worker (e.g. c("data.table) if data stored as such)}

\item{...}{Additonal arguments passed to \code{gamlss()}.}
}
\value{
A cumulative density function
}
\description{
Fit a gamlss paramertirc forecast model
}
\details{
Details go here...
}
