% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Para_gamlss.R
\name{Para_gamlss}
\alias{Para_gamlss}
\title{Fit \code{gamlss}-type semi-paramertirc models}
\usage{
Para_gamlss(
  data,
  formula,
  sigma.formula = ~1,
  nu.formula = ~1,
  tau.formula = ~1,
  family = NO(),
  parallel = F,
  cores = NULL,
  pckgs = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory variables.
May optionally contain a collumn called "\code{kfold}" with numbered/labeled folds
and "\code{Test}" for test data.}

\item{formula}{A formula object with the response on the left of an ~ operator,
and the terms, separated by + operators, on the right.}

\item{sigma.formula}{A formula object for fitting a model to the
sigma parameter, as in the formula above.}

\item{nu.formula}{A formula object for fitting a model to the
nu parameter, as in the formula above.}

\item{tau.formula}{A formula object for fitting a model to the
tau parameter, as in the formula above.}

\item{family}{A gamlss.family object, which is used to define the
distribution and the link functions of the various parameters.}

\item{parallel}{\code{boolean} parallelize model fitting process? Parallelisation is
over cross-validation folds.}

\item{cores}{if parallel is TRUE then number of available cores}

\item{pckgs}{if parallel is TRUE then  specify packages required
for each worker (e.g. c("data.table) if data stored as such)}

\item{...}{Additonal arguments passed to \code{gamlss()}.}
}
\value{
A list of \code{gamlss} models with class \code{PPD}. Each list element
corresponsds to a cross-validation fold and contains a \code{gamlss} model
trained on all other folds.
}
\description{
This function is a wrapper for the function \code{gamlss}, which
fits semi-parametric regression models for predictive disributions with up to
four parameters (location, scale, shape1, shape2).
}
\details{
See \code{?gamlss} for additional details and options.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}; Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
