% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinball.R
\name{pinball}
\alias{pinball}
\title{Pinball Loss for MultiQR}
\usage{
pinball(
  qrdata,
  realisations,
  kfolds = NULL,
  plot.it = T,
  subsets = NULL,
  breaks = 4,
  bootstrap = NULL,
  ...
)
}
\arguments{
\item{qrdata}{\code{MultiQR} object.}

\item{realisations}{Vector of realisations corresponding to rows of \code{qrdata}. \code{NA} accepted.}

\item{kfolds}{Optional vector of fold/test labels corresponding to rows of \code{qrdata}.}

\item{plot.it}{\code{boolean}. Make a plot?}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
Quantile Score data and, if plot.it=T, a reliability diagram.
}
\description{
This function calculates and plots the pinball loss for a MultiQR object and plots the result
}
\details{
Details go here...
}
