% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIT.R
\name{PIT.MultiQR}
\alias{PIT.MultiQR}
\title{Probability integral transform: S3 Method for for MultiQR}
\usage{
\method{PIT}{MultiQR}(qrdata, obs, tails, inverse = FALSE, ...)
}
\arguments{
\item{qrdata}{A \code{MultiQR} object.}

\item{obs}{A vector of observations corresponding to the rows of \code{qrdata}.}

\item{tails}{A list of arguments passed to \code{condCDF} defining the tails of the CDFs.}

\item{inverse}{A \code{boolean}. If true, the inverse transformation is appiled, i.e. uniform variable
to random variable from original distribution.}

\item{...}{Additional arguments passed to \code{condCDF}.}
}
\value{
The probability integral transform of \code{obs} through the
predictive distribution defined by \code{qrdata} and
interpolation scheme in \code{contCDF} with \code{tails}.
}
\description{
Transforms data (random variables) through their corresponding cumulative
distribution function (specified by a \code{MultiQR} object and tail parameters) in 
order to produce variables with a standard uniform distribution.
}
\details{
Boundary imposed throwing a warning to ensure that output is in [0,1].
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
