% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRfunctions.R
\name{PIT.MultiQR}
\alias{PIT.MultiQR}
\title{Probability integral transform: S3 Method for for MultiQR}
\usage{
\method{PIT}{MultiQR}(qrdata, obs, tails, ...)
}
\arguments{
\item{qrdata}{A \code{MultiQR} object.}

\item{obs}{A vector of observations corresponding to \code{qrdata}.}

\item{tails}{A list of arguments passed to \code{condCDF} defining the tails of the CDF}

\item{...}{Additional arguments passed to \code{condCDF}.}
}
\value{
The probability integral transform of \code{obs} through the predictive distribution defined by \code{qrdata} and interpolation scheme in \code{contCDF}.
}
\description{
This function produces a fan plot of a MultiQR object.
}
\details{
Details go here...
}
