% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIT.R
\name{PIT.PPD}
\alias{PIT.PPD}
\title{Probability integral transform: S3 Method for for PPD}
\usage{
\method{PIT}{PPD}(ppd, data, inverse = FALSE, inv_probs = NULL)
}
\arguments{
\item{ppd}{A \code{PPD} object (a list of \code{gamlss} objects).}

\item{data}{Input data corresponding to \code{ppd}.}

\item{inverse}{A \code{boolean}. If true, the inverse transformation is appiled,
i.e. uniform variable.}

\item{inv_probs}{If \code{inverse}, vector of uniform-distributed variable
to be transformed.}
}
\value{
The probability integral transform of \code{data} through
the predictive distribution defined by \code{PPD}.
}
\description{
Transforms data (random variables) through their corresponding cumulative
distribution function (specified by a \code{PPD} object) in 
order to produce variables with a standard uniform distribution.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
