% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRfunctions.R
\name{plot.MultiQR}
\alias{plot.MultiQR}
\title{Plot a Quantile Forecast}
\usage{
\method{plot}{MultiQR}(plotdata, targetTimes = NULL,
  quantiles = colnames(plotdata), ...)
}
\arguments{
\item{plotdata}{A \code{MultiQR} object to be plotted}

\item{targetTimes}{A vector of forecast target times corresponding to \code{plotdata}.}

\item{quantiles}{A charater vector. By default the column names from MultiQR object, i.e. all available quantiles. Alternatively, a subset may be specified here.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
A plot of a \code{MQR}.
}
\description{
This function produces a fan plot of a MultiQR object.
}
\details{
Details go here...
}
\keyword{Quantile}
\keyword{Regression,}
\keyword{plot}
