% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_gam.R
\name{predict.qreg_gam}
\alias{predict.qreg_gam}
\title{Predict from model based on Generalised Additive Model and Linear Quantile Regression}
\usage{
\method{predict}{qreg_gam}(
  object,
  newdata = NULL,
  quantiles = NULL,
  cv_fold = NULL,
  sort = T,
  sort_limits = NULL,
  ...
)
}
\description{
This function predicts from multiple conditional linear quantile regression models of the residuals of
a generalised additive model fit using \code{mqr_qreg_gam}.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
