% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_gbm.R
\name{predict.qreg_gbm}
\alias{predict.qreg_gbm}
\title{Predict method for Multiple Quantile Regression from Gradient Boosted Decision Trees.}
\usage{
\method{predict}{qreg_gbm}(
  object,
  newdata = NULL,
  quantiles = NULL,
  pred_ntree = NULL,
  perf.plot = FALSE,
  sort = T,
  sort_limits = NULL,
  ...
)
}
\arguments{
\item{object}{object of class \code{qreg_gbm} obtained from the function \code{qreg_gbm()}}

\item{newdata}{data.frame of observations for which to make predictions}

\item{quantiles}{The quantiles to predict. Default is all the quantiles present in \code{object}}

\item{pred_ntree}{predict using a user-specified tree.
If unspecified an out-of-the bag estimate will be used unless internal
gbm cross-validation folds are specified in \code{qreg_gbm()}}

\item{perf.plot}{plot GBM performance if \code{pred_ntree = NULL}?}

\item{sort}{sort quantiles using \code{SortQuantiles()}?}

\item{sort_limits}{\code{Limits} argument to be passed to \code{SortQuantiles()}. Constrains quantiles to upper and 
lower limits given by \code{list(U=upperlim,L=lowerlim)}.}

\item{...}{additional arguments; not currently used.}
}
\value{
Quantile forecasts in a \code{MultiQR} object.
}
\description{
This function returns multiple quantile predictions as an object of class \code{MultiQR}
based on a ProbCast gbm fitted model. S3 Method for for \code{qreg_gbm} objects
}
\details{
this function returns predictive quantiles for each row in \code{newdata},
the result is returned as a \code{MultiQR} object
}
\author{
Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
