% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_gam.R
\name{qreg_gam}
\alias{qreg_gam}
\title{Multiple Quantile Regression Using Generalised Additive Models and Linear Quantile Regression}
\usage{
qreg_gam(
  data,
  formula,
  formula_qr = NULL,
  model_res2 = F,
  formula_res2 = formula,
  quantiles = c(0.25, 0.5, 0.75),
  cv_folds = NULL,
  use_bam = T,
  sort = T,
  sort_limits = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory
variables. May optionally contain a collumn called "kfold" with
numbered/labeled folds and "Test" for test data. If \code{data} contains a column called
"gam_pred" then gam modelling will be skipped and this will be used in quantile regression.}

\item{formula_qr}{Formula for linear quantile regression model for GAM residuals. Term \code{gam_pred}
is the prediction from the above GAM may be included in this formula. If null, the "terms" of the GAM model
are used as features in linear quantile regression.}

\item{model_res2}{If \code{TRUE} also model squared residuals of GAM using a GAM. Defaults to \code{FALSE}.}

\item{formula_res2}{Formula for GAM to predict squared residuals.}

\item{quantiles}{The quantiles to fit models for.}

\item{cv_folds}{Control for cross-validation if not supplied in \code{data}.}

\item{use_bam}{If \code{TRUE} (default) then GAM is fit using (\code{bam()}) in stead of \code{gam()}. \code{bam} is better suited to large datasets but not all \code{gam} model options are available with \code{bam}. See \code{bam()} documentation for further details.}

\item{...}{Additional agruments passter to \code{gam()} or (\code{bam()}).}

\item{formala}{A \code{formula} object with the response on the left
of an ~ operator, and the terms, separated by + operators, on the right passed to \code{gam()}
or \code{bam()} from \code{mgcv}.}

\item{gbm_params}{List of parameters to be passed to \code{fit.gbm()}.}

\item{Sort}{\code{boolean} Sort quantiles using \code{SortQuantiles()}?}

\item{w}{Weights on the contribution of data to model fit. See \code{gam()}.}

\item{SortLimits}{\code{Limits} argument to be passed to \code{SortQuantiles()}. Constrains quantiles to upper and lower limits given by \code{list(U=upperlim,L=lowerlim)}.}
}
\value{
Returns a \code{list} containing predictive quantiles (in a \code{MultiQR} object), GAM models, and deterministic predictions
from GAMs.
}
\description{
This function fits multiple conditional linear quantile regression models to the residuals of
a generalised additive model using \code{mgcv} and with facilities for cross-validation.
}
\details{
The returned predictive quantiles and GAM predictions are those produced out-of-sample for each
cross-validation fold (using models trained on the remaining folds but not "Test" data).
Predictive quantiles corresponding to "Test" data are produced using models trained on all
non-test data.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
