% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQR_qreg_mboost.R
\name{qreg_mboost}
\alias{qreg_mboost}
\title{Multiple Quantile Regression using \code{mboost}}
\usage{
qreg_mboost(
  data,
  formula,
  quantiles = c(0.25, 0.5, 0.75),
  cv_folds = NULL,
  w = rep(1, nrow(data)),
  cores = 1,
  pckgs = NULL,
  sort = T,
  sort_limits = NULL,
  save_models_path = NULL,
  only_mqr = FALSE,
  exclude_train = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing target and explanatory variables.}

\item{quantiles}{The quantiles to fit models for.}

\item{cv_folds}{Control for cross-validation with various options, either:
\itemize{
 \item the column name of the fold index supplied in data. Observations and inputs 
  in the index labelled "Test" will serve as test data and held out in model training.
 \item an integer giving the number of cross validation folds to generate. Folds are constructed as block chunks. 
 Default behaviour is 5 folds.
 \item vector of length==nrow(data) containing character or numeric fold labels.
 \item NULL indicates that no cross validation should be performed and the returned model is trained on all \code{data}.
}}

\item{w}{an optional numeric vector of weights to be used in the fitting process.}

\item{cores}{the number of available cores. Defaults to one, i.e. no parallelisation, although in 
this case the user must still specify \code{pckgs} if applicable.}

\item{pckgs}{specify additional packages required for
each worker (e.g. c("data.table") if data stored as such).}

\item{sort}{Sort quantiles using \code{SortQuantiles()}?}

\item{sort_limits}{\code{Limits} argument to be passed to \code{SortQuantiles()}. Constrains quantiles 
to upper and lower limits given by \code{list(U=upperlim,L=lowerlim)}.}

\item{save_models_path}{Path to save models. Model details and file extension pasted onto this string.
Defaults to \code{NULL}, i.e. no model save.}

\item{only_mqr}{return only the out-of-sample predictions?}

\item{exclude_train}{control for exclusion of rows in data for the model training only, with various options, either:
\itemize{
 \item the column name of the binary/boolean exclude flag supplied in data.
 \item a vector of binary/boolean exclusion flags of length nrow(data)
 \item NULL indicates no exclusion
}
This option is useful when out-of-sample predictions are required in rows which need excluded during model training}

\item{...}{extra hyper-parameters to be passed to \code{mboost()}.
e.g. use \code{control = mboost::boost_control()} to specify boosting steps and shrinkage.}

\item{formaula}{A \code{formula} object with the response on the left
of an ~ operator, and the terms, separated by + operators, on the right.}
}
\value{
by default a named list containing fitted models as a list of \code{qreg_mboost} objects, 
and out-of-sample cross validation  forecasts as an \code{MultiQR} object. The output list depends on \code{cv_folds}.

Alternatively returns only the out-of-sample cross validation forecasts as an \code{MultiQR} 
object when \code{only_mqr} is \code{TRUE}
}
\description{
This function fits multiple quantile regression models using \code{mboost}, with
facilities for cross-validation. \code{mboost} accommodates both generalised
additive models, decision trees and other learners. See \code{?mboost} for more
details.
}
\details{
The returned predictive quantiles are those produced out-of-sample for each
cross-validation fold (using models trained on the remaining folds but not "Test" data).
Predictive quantiles corresponding to "Test" data are produced using models trained on all
non-test data.

The returned models are in a named list corresponding to the model for each fold and 
and can be extracted for further prediction or evaluation. See \code{predict.qreg_mboost()}.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}; Ciaran Gilbert, \email{ciaran.gilbert@strath.ac.uk}
}
\keyword{Quantile}
\keyword{Regression}
