% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateFeatures.R
\name{read_holiday_ics}
\alias{read_holiday_ics}
\title{Load data of UK Holidays}
\usage{
read_holiday_ics(folder = getwd())
}
\arguments{
\item{folder}{Folder location of .ics file(s) of UK holidays
from https://www.gov.uk/bank-holidays}
}
\value{
Table of UK Holidays
}
\description{
Function to read official data of UK public holidays.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
\keyword{Date}
\keyword{Features}
