% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Reliability Diagram for MultiQR}
\usage{
reliability(qrdata, realisations, kfolds = NULL, plot.it = T, ...)
}
\arguments{
\item{qrdata}{\code{MultiQR} object.}

\item{realisations}{Vector of realisations corresponding to rows of \code{qrdata}. \code{NA} accepted.}

\item{kfolds}{Optional vector of fold/test labels corresponding to rows of \code{qrdata}.}

\item{plot.it}{\code{boolean}. Make a plot?}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
Reliability data and, if plot.it=T, a reliability diagram.
}
\description{
This function plots a reliabiltiy diagram for a MultiQR object and returns the plotdata
}
\details{
Details go here...
}
