% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Reliability Diagram for MultiQR}
\usage{
reliability(
  qrdata,
  realisations,
  kfolds = NULL,
  subsets = NULL,
  breaks = 4,
  bootstrap = NULL,
  plot.it = T,
  ...
)
}
\arguments{
\item{qrdata}{\code{MultiQR} object.}

\item{realisations}{Vector of realisations corresponding to rows of \code{qrdata}. \code{NA} accepted.}

\item{kfolds}{Optional vector of fold/test labels corresponding to rows of \code{qrdata}. Cannot be used with \code{subsets}.}

\item{subsets}{Optional vector of covariates to bin data by. Breaks between bins are the empirical quantiles of \code{subsets}. Cannot be used with \code{kfold}.}

\item{breaks}{Number of quantiles to divide subsets by, results in \code{breaks+1} bins.}

\item{bootstrap}{Calculate this number of boostrap samples to estimate 95\% confdence interval}

\item{plot.it}{\code{boolean}. Make a plot?}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
Reliability data and, if \code{plot.it=TRUE}, a reliability diagram.
}
\description{
This function plots a reliabiltiy diagram for a MultiQR object and returns the plotdata
}
\details{
Details go here...
}
