% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Reliability Diagram for MultiQR}
\usage{
reliability(
  qrdata,
  realisations,
  kfolds = NULL,
  subsets = NULL,
  breaks = 4,
  bootstrap = NULL,
  plot.it = T,
  ...
)
}
\arguments{
\item{qrdata}{\code{MultiQR} object.}

\item{realisations}{Vector of realisations corresponding to
rows of \code{qrdata}. Missing data as \code{NA}s accepted.}

\item{kfolds}{Optional vector of cross-validation fold labels corresponding
to rows of \code{qrdata}. Cannot be used with \code{subsets}.}

\item{subsets}{Optional vector of covariates to bin data by.
Breaks between bins are the empirical quantiles of
\code{subsets} bu default or all unique factors or charater strings.
Cannot be used with \code{kfolds}.}

\item{breaks}{Number of quantiles to divide subsets by, results
in \code{breaks+1} bins, or is length > 1, a vector of break points.}

\item{bootstrap}{Calculate this number of boostrap samples
to estimate 95\% confdence interval}

\item{plot.it}{\code{boolean}. Make a plot?}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
Reliability data and, if \code{plot.it=TRUE}, a reliability diagram.
}
\description{
Calculated empirical exceedence of each quantile and 
plots a reliabiltiy diagram for a \code{MultiQR} object.

Optionally, results may be split by cross-validation fold or covariate and/or
confidence intervals may be estimated.
}
\details{
Missing values in \code{realisations} are handled by \code{na.rm=T} when
calculating average exceedence of a given quantile.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
