% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleMQR.R
\name{rescaleMQR}
\alias{rescaleMQR}
\title{Re-scale MultiQR object}
\usage{
rescaleMQR(mqr, mult = 1, add = 0, Order = "ma")
}
\arguments{
\item{mqr}{A \code{MultiQR} object to be re-scaled}

\item{mult}{Multiplicative adjustment of length 1 or with length equal
to the number of rows in mqr.}

\item{add}{Additive adjustment of length 1 or with
length equal to the number of rows in mqr.}

\item{order}{Order in which to perform adjustment, either
\code{"ma"} (multiple-then-add) or vice versa.}
}
\value{
A \code{MultiQR} object.
}
\description{
This function re-scales MultiQR object.
}
\details{
This function re-scales \code{MultiQR} objects. Typical usage
is to return forecasts in the domain of the original target variable if modelling
and forecasting has been performed after normalisation/standardisation.
}
\keyword{Quantile}
\keyword{Regression,}
\keyword{plot}
