% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRfunctions.R
\name{samps_to_scens}
\alias{samps_to_scens}
\title{Generate multivariate forecasts}
\usage{
samps_to_scens(copulatype, no_samps, list_margins, list_sigma, list_mean,
  control, ...)
}
\arguments{
\item{copulatype}{Either "spatial" of "temporal", note that spatio-temporal can be generated via "temporal" setting}

\item{no_samps}{Number of scenarios required}
}
\value{
A list or data frame of multivariate scenario forecasts
}
\description{
This function produces a list of multivariate scenario forecasts in the marginal domain from the spatial/tempral/spatiotemporal covariance matrices and marginal distributions
}
\details{
Details go here...
}
